Move hooks into one place
